pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[project]
name = "email-batch-app"
version = "0.1.0"
description = "FastAPI app for personalized Gmail batch sending"
authors = [{name = "Codex"}]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.3",
    "itsdangerous>=2.1.2",
    "pydantic[email]>=2.6.4",
    "pydantic-settings>=2.2.1",
    "python-docx>=1.1.0",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.120.0",
    "cryptography>=42.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "black>=24.2.0",
    "isort>=5.13.0",
    "mypy>=1.8.0"
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.setuptools]
packages = ["app"]
include-package-data = true
